toolchain_build_pnacl.py: Build newlib
Build newlib, the first of the PNaCl target bitcode libraries.
Because the toolchain components are not assembled together incrementally as
as they were with build.sh, use a copy of the driver scripts with a driver.conf
that points to the host binaries to build the target libs.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3719

Review URL: https://codereview.chromium.org/138063010

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12628 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2