Add support for detach and kill to the debug stub.
Allow the debug stub to leave threads in a suspended state if a detach occurs.
Correctly interpret 'D' as a detach (rather than 'd' which I assume was a typo).
Disconnect is also interpreted as a detach.

Kill causes sel_ldr to exit with 247 (-9).

Select the FD_CLOSE event on windows to allow proper disconnect handling.

R=mseaborn@chromium.org, bsy@google.com
BUG=None
TEST=debug stub tests (added detach, kill, complete, disconnect)

Review URL: https://codereview.chromium.org/105073008

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12629 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2