Updated Once so that it remembers the list of all cached directories.
In the past, the toolchain build memoization step in Once just merely
used directory_storage to upload/download directories and passed off the
URL to a simple print function. The directory_storage module has been
updated to return a named tuple (DirectoryStorageItem) upon a successful
upload/download which contains a bit more information about the directory
(name, hash of the tar, and the url). Memoized directories are now kept
track of in a list and can be queried later.

R=bradnelson@google.com, mcgrathr@chromium.org
BUG=
TEST= trybots

Review URL: https://codereview.chromium.org/184233011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12845 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2