Split NaClChromeMainStartApp() for Chromium IPC.
To start NaCl applications over Chromium IPC, we have to split up
NaClChromeMainStartApp() so that load failures may be reported to the renderer.

NaClChromeMainStartApp() is split into two functions:
NaClChromeMainLoad(), which loads the IRT and the user nexe, and currently blocks until
START_MODULE is sent over SRPC.

NaClChromeMainStart(), which starts the NaCl application thread and waits for
it to exit.

BUG=333950
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/195733011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12912 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2