Added back nacl_glibc_tc_root variable in untrusted.gypi
Previously the nacl_glibc_tc_root variable had been removed because
it was pointing to the wrong toolchain location and was not used.
However it is actually being used on the chromium side gyp commands.
I have added it back it now, and also corrected the location to point
to the correct gyp toolchain directory.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3816
TEST= trybots
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/206703002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12922 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2