PNaCl/NaCl test: add support for symbol wrapping through -Wl,--wrap
Support was removed a while ago because it was only used in the
now-deprecated nacl-mounts, otherwise completely untested, and based on
local modifications to Gold and LLVM.

Symbol wrapping is nonetheless used by some SDK users, and works with
GCC and regular LLVM, it would therefore be nice for it to work with
PNaCl.

The reverts, from which I cherry-picked the symbol wrapping code and
rebased to the latest Gold and LLVM:
https://codereview.chromium.org/22831032
https://codereview.chromium.org/119203005
https://codereview.chromium.org/23398002

The latest bug in which this code was changed (before deletion):
https://code.google.com/p/nativeclient/issues/detail?id=2359

For reference, nacl_io wraps Newlib APIs in a completely different way:
https://chromium.googlesource.com/experimental/chromium/src/+/refs/heads/master/native_client_sdk/src/libraries/nacl_io/kernel_wrap_newlib.cc

This is a four part change:
  LLVM: https://codereview.chromium.org/203623007
  Gold: https://codereview.chromium.org/197883026
  NaCl test: https://codereview.chromium.org/196943027
  pnacl-ld: https://codereview.chromium.org/205153004

This CL also update TOOL_REVISIONS for PNaCl r12914->r12917.

Pull the following PNaCl changes into NaCl:
  r12916: (jfb@chromium.org) PNaCl: Update LLVM revision in pnacl/COMPONENT_REVISIONS
    | 17d8913: (kschimpf@google.com) Moves abbreviation size computations to abbreviation blocks.
    | 1b2d7a6: (jfb@chromium.org) PNaCl/LLVM: add support for symbol wrapping through -Wl,--wrap
  r12917: (jfb@chromium.org) PNaCl: Update Binutils revision in pnacl/COMPONENT_REVISIONS
    | b7670a7: (jfb@chromium.org) PNaCl/Gold: add support for symbol wrapping through -Wl,--wrap

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3818
TEST= ./scons run_wrap_test bitcode=0 && ./scons run_wrap_test bitcode=1
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/196943027

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12924 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2