toolchain_build should only upload what we will download for (arm).
There was previous discussions regarding toolchain_build uploading for
all target architectures it builds for and only having the downloader
choose what and when to download. However I realized nacl_x86_newlib is
being uploaded by another builder (the makefile) so uploading that
can potentially stomp it over. For this reason, we should keep track
of a separate list (UPLOAD_TARGETS) which keep track of what targets
we want to upload for.

BUG=
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/256433004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13078 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2