Do not instument nacl_bootstrap_munge_phdr with MemorySanitizer.
This target is built with -lelf, so we would have to provide MSan-instrumented
libelf to be able to run it correctly. Since this small helper utility is the
only target in Chromium that depends on libelf, it's easier to just ignore it.

Also, strip -stdlib=libc++ for nacl_bootstrap_munge_phdr and nacl_bootstrap_lib.

BUG=chromium:362159, chromium:358678
R=mseaborn@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/225163009

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13145 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2