Git repositories syncing through repo_tools can be validated.
In the cases of invalid local git repositories and updated URLs it is
useful to be able to clobber the git repository if it detects that it
is not configured as expected. Since clobbering is potentially scary
I have made this option a function argument that defaults to False.

I have also turned on this argument in toolchain_build as I will be
changing the git URLs in a later changelist. This will allow us to roll
back any changes easily.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3852
TEST= toolchain_build trybots
R=dschuff@chromium.org, mcgrathr@chromium.org

Review URL: https://codereview.chromium.org/284763006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13215 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2