toolchain_build: Fix arm-linux-gnueabi-gcc -v collection
The change in r13265 had a typo that meant it used 'gcc -v'
output in place of 'arm-linux-gnueabi-gcc -v' and so did not have
the effect intended.  Fix it by factoring out cut&pasted code
into a subroutine, so there cannot be a cut&paste editting error.

BUG= none
TEST= trybots
R=bradnelson@google.com

Review URL: https://codereview.chromium.org/317433002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13270 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2