When validating git repositories, also check if the repository is empty.
During the course of testing git repositories, I accidentally filled
some git repositories with invalid data. This has triggered some errors
when cloning because the repositories need to be empty before they can
be cloned into. It is probably a good idea for the validation pass to
check for this.

R=bradnelson@google.com
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3852
TEST= toolchain_build trybots

Review URL: https://codereview.chromium.org/317803008

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13288 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2