Set the default number of threads to be 10.
The motivation originally given to set this to 1 is that mac was running out of processes. maxprocs is now 2000 (from 1300).

On Mac, this speeds up building nacl_irt from 13 to 4 minutes.

Goma is disabled when building NaCl on Windows because of bug 359729.

This is a duplicate of https://codereview.chromium.org/358533002/, trying again from git the first time since it wasn't patching.

BUG=388035
NOTRY=true

Review URL: https://codereview.chromium.org/357773004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13407 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2