PNaCl dynamic loading: Extend pnacl-translate to be able to translate PSOs
Add a "-pso" option to tell pnacl-translate to treat the input as a PSO
and translate it to an ELF DSO.

In principle, pnacl-translate could auto-detect the difference between a
pexe and a PSO, but that would involve opening the file to check which
symbols it defines, which might slow down pnacl-translate unnecessarily.

The minimal test for this just checks that the DSO is created
successfully.  I've also tested this against the PNaCl dynamic loader
implementation that I'll add in a follow-on change.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3873
TEST=run_pnacl_dynamic_loading_test
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/352213002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13413 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2