Stub out pnacl-opt and pnacl-finalize tools for scons built_elsewhere=1
This is required to make the PSO generation test pass on ARM hardware;
(or more specifically, opt will run on the x86 bot, and get stubbed
out on the hardware bot). Stubbing out pnacl-finalize isn't
required for that particular test but is just for completeness.
R=mseaborn@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3873

Review URL: https://codereview.chromium.org/388573005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13469 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2