Add embedder interface for fatal log messages.
For Chromium, we'd like to report fatal log messages to the JavaScript console
without using SRPC. This change introduces an EmbedderInterface which will
contain callbacks for some of the plugin->renderer messages so that the
embedder can use whatever transport it likes.

This follows the pattern used by NaClValidationCache, which serves a similar
purpose.

BUG=chromium:391039
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/367023004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13523 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2