Followup to "do not generate temp file names beyond Windows' MAX_PATH"
Don't normalize the path before counting the length.
It appears that tools like clang won't normalize the path,
and that counts against the limit.

An alternative is to go through each of the inputs and pass
the normalized absolute path to the tools, instead of the relative paths.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3846
R=dschuff@chromium.org

Review URL: https://codereview.chromium.org/420713007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@13539 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2